requirements.txt
configparser==4.0.2
hexdump==3.3
more-itertools==5.0.0
pyparsing==2.4.7
python-evtx==0.7.4
six==1.16.0
zipp==1.0.0
